Conflicts:
	App/src/FitSuiteHelper.cpp
	App/src/IsGISAXSTools.cpp
	App/src/TestFittingModule3.cpp
	Core/Tools/inc/FitSuiteObjects.h

+++ App/inc/IsGISAXSTools.h
@@ -29,7 +29,7 @@ class TLine;
 class IsGISAXSTools {
 public:
     struct AxisStructure {
-        int nbins;
+        size_t nbins;
         std::vector<double> xbins;
         std::string name;
     };
